Start was creating the page via .jsp files. The second step was to validate the input.
Next was creating the backend for the remember me button using cookies
to populate the input box for the username. Deleting the cookies stops
this from happening.

I used a session to check if the user had logged in, and keep them
logged in if they entered the right username and password when you
change the URL to point to the login page.

You cannot go to the home page when the session has been invalidated.