Differentiating between forwards and redirecting is of utmost importance for correct web site behavior. If the user enters the correct username and password they are redirected to the home page.
If the username and pass are incorrect an error message is thrown.

If the user is logged in the session stops them from going to the login
page if the session is not null. If the logout button is pressed the
user is redirected to the login page.

You cannot go to the home page without signing in properly and you
cannot go to the login page with a valid session.

If the remember me check box is checked and the user logs in properly
the username is entered as a cookie in the browser. If the check box
is not checked the cookie is deleted.